+++ aldrin/CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   new functions `Reply::timestamp()` and `low_level::Reply::timestamp()`.
 - Add type `Property` to help keeping track of remote values.
 - Add `low_level::PendingReply::id()`.
+- Add `low_level::Call::take_args()`.
 
 ### Changed